Refactoring to packages. Closes #48
