+++ tests/unit/test_crypto.py
@@ -18,7 +18,7 @@ import base64
 import pytest
 
 from pysui.abstracts.client_keypair import SignatureScheme
-from pysui.sui.sui_types import address_from_keystring
+from pysui.sui.sui_types.address import address_from_keystring
 from pysui.sui.sui_crypto import (
     SuiPublicKeyED25519,
     SuiPrivateKeyED25519,
